Move HTTP Proxy and remove and relocate duplicate code
