+++ include/Http.h
@@ -14,7 +14,7 @@ public:
     static std::string Post(const std::string& IP, const std::string& Fields);
     static std::string Get(const std::string& IP);
     static bool ProgressBar(size_t c, size_t t);
-
+    static void StartProxy();
 public:
     static bool isDownload;
 };
