+++ include/Startup.h
@@ -14,8 +14,8 @@ void InitLauncher(int argc, char* argv[]);
 std::string GetEP(char* P = nullptr);
 std::string GetGamePath();
 std::string GetVer();
+std::string GetPatch();
 std::string GetEN();
-void StartProxy();
 void ConfigInit();
 extern bool Dev;
