+++ include/Utils.h
@@ -1,0 +1,20 @@
+#pragma once
+#include <string>
+#include <vector>
+
+namespace Utils {
+    inline std::vector<std::string> Split(const std::string& String, const std::string& delimiter) {
+        std::vector<std::string> Val;
+        size_t pos;
+        std::string token, s = String;
+        while ((pos = s.find(delimiter)) != std::string::npos) {
+            token = s.substr(0, pos);
+            if (!token.empty())
+                Val.push_back(token);
+            s.erase(0, pos + delimiter.length());
+        }
+        if (!s.empty())
+            Val.push_back(s);
+        return Val;
+    };
+};
